requirements (1).txt

requirements = '''streamlit
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.13
chromadb==0.4.22
pypdf==3.17.4
pandas
plotly
pyyaml'''
